Scoverage added + general refactoring
Currently, with all generators and BaseView[^5] excluded coverage is 48%.

Dependencies updated; Scala 2.11.4 -> 2.11.5; as support for 2.10 was dropped
I removed it from build scripts.

Imports optimized, empty comments and author comments removed.

All generators (not generated code) moved to `utils.generators` package to
enable easy exclusion from coverage.
