+++ .travis.yml
@@ -1,8 +1,8 @@
 language: scala
 scala:
-  - "2.10.4"
-  - "2.11.2"
+  - "2.11.5"
 jdk:
   - openjdk7
   - oraclejdk8
-sbt_args: -no-colors
+
+script: sbt -no-colors ++$TRAVIS_SCALA_VERSION clean coverage test
