+++ src/test/scala/org/virtuslab/beholder/FormFiltersTests.scala
@@ -2,16 +2,13 @@ package org.virtuslab.beholder
 
 import java.sql.Date
 
-import org.virtuslab.beholder.filters.{ FilterAPI, FilterDefinition }
 import org.virtuslab.beholder.filters.forms.FromFilterFields._
 import org.virtuslab.beholder.filters.forms.{ FormFilters, FormFormatter, FromFilterFields }
+import org.virtuslab.beholder.filters.{ FilterAPI, FilterDefinition }
 import org.virtuslab.beholder.suites.{ BaseSuite, FiltersTestSuite, RangeFiltersSuite }
 import org.virtuslab.unicorn.LongUnicornPlay._
 import org.virtuslab.unicorn.LongUnicornPlay.driver.simple._
 
-/**
- * Author: Krzysztof Romanowski
- */
 trait FormFiltersTestsBase {
   self: AppTest with BaseSuite[FormFormatter] =>
