+++ src/test/scala/org/virtuslab/beholder/model/User.scala
@@ -2,6 +2,7 @@ package org.virtuslab.beholder.model
 
 import org.virtuslab.unicorn.LongUnicornPlay._
 import org.virtuslab.unicorn.LongUnicornPlay.driver.simple._
+
 import scala.slick.lifted.Tag
 
 /** Id class for type-safe joins and queries. */
