+++ src/test/scala/org/virtuslab/beholder/model/UserMachines.scala
@@ -1,6 +1,7 @@
 package org.virtuslab.beholder.model
 
 import org.virtuslab.unicorn.LongUnicornPlay.driver.simple._
+
 import scala.slick.lifted.{ TableQuery, Tag }
 
 class UserMachines(tag: Tag) extends Table[(UserId, MachineId)](tag, "users_machines") {
