+++ src/test/scala/org/virtuslab/beholder/repositories/MachineRepository.scala
@@ -1,8 +1,9 @@
 package org.virtuslab.beholder.repositories
 
-import org.virtuslab.beholder.model.{ Machines, Machine, MachineId }
+import org.virtuslab.beholder.model.{ Machine, MachineId, Machines }
 import org.virtuslab.unicorn.LongUnicornPlay._
 import org.virtuslab.unicorn.LongUnicornPlay.driver.simple._
+
 import scala.slick.lifted.TableQuery
 
 /**
